+++ indexing-service/src/main/java/org/apache/druid/indexing/common/task/batch/parallel/PartialRangeSegmentGenerateTask.java
@@ -196,7 +196,7 @@ public class PartialRangeSegmentGenerateTask extends PartialSegmentGenerateTask<
       TaskToolbox toolbox,
       List<DataSegment> segments,
       Map<String, TaskReport> taskReport,
-      Integer segmentsRead
+      Long segmentsRead
   )
   {
     List<PartitionStat> partitionStats = segments.stream()
